requirements.txt
Django==1.5.8
Fabric==1.8.3
South==0.8.4
amqp==1.4.5
anyjson==0.3.3
billiard==3.3.0.17
boto==2.29.1
celery==3.1.11
dj-database-url==0.1.2
django-allauth==0.16.1
django-analytical==0.18.0
django-annoying==0.8.0
django-appconf==0.6
django-celery==3.1.10
django-celery-with-redis==3.0
django-compressor==1.4
django-debug-toolbar==1.2.1
django-extensions==1.3.7
django-heroku-memcacheify==0.5
django-heroku-postgresify==0.3
django-http-proxy==0.3.2
django-nose==1.2
django-pylibmc-sasl==0.2.4
django-ses==0.6.0
django-staticfiles==1.2.1
django-storages==1.1.8
ecdsa==0.11
gunicorn==18.0
kombu==3.0.18
nose==1.3.3
oauthlib==0.6.1
paramiko==1.12.4
psycopg2==2.5.3
pycrypto==2.6.1
pylibmc==1.2.3
python-openid==2.2.5
pytz==2014.4
redis==2.10.1
requests==2.3.0
requests-oauthlib==0.4.0
six==1.6.1
sqlparse==0.1.11
wsgiref==0.1.2
